+++ notifications_utils/pdf.py
@@ -1,8 +1,8 @@
 import io
 
 import PyPDF2
-from PyPDF2 import PdfFileWriter
-from PyPDF2.utils import PdfReadError
+from PyPDF2 import PdfWriter
+from PyPDF2.errors import PdfReadError
 
 from notifications_utils import LETTER_MAX_PAGE_COUNT
 
@@ -11,14 +11,14 @@ def pdf_page_count(src_pdf):
     """
     Returns number of pages in a pdf file
 
-    :param PyPDF2.PdfFileReader src_pdf: A File object or an object that supports the standard read and seek methods
+    :param PyPDF2.PdfReader src_pdf: A File object or an object that supports the standard read and seek methods
     """
     try:
-        pdf = PyPDF2.PdfFileReader(src_pdf)
+        pdf = PyPDF2.PdfReader(src_pdf)
     except AttributeError as e:
         raise PdfReadError("Could not open PDF file, stream is null", e)
 
-    return pdf.numPages
+    return len(pdf.pages)
 
 
 def is_letter_too_long(page_count):
@@ -38,16 +38,16 @@ def extract_page_from_pdf(src_pdf, page_number):
     :param src_pdf: File object or an object that supports the standard read and seek methods similar to a File object.
     :param page_number: The page number to retrieve (pages begin at zero)
     """
-    pdf = PyPDF2.PdfFileReader(src_pdf)
+    pdf = PyPDF2.PdfReader(src_pdf)
 
-    if pdf.numPages < page_number:
+    if len(pdf.pages) < page_number:
         raise PdfReadError("Page number requested: {} of {} does not exist in document".format(
             str(page_number),
-            str(pdf.numPages)
+            str(len(pdf.pages))
         ))
 
-    writer = PdfFileWriter()
-    writer.addPage(pdf.getPage(page_number))
+    writer = PdfWriter()
+    writer.add_page(pdf.pages[page_number])
 
     pdf_bytes = io.BytesIO()
     writer.write(pdf_bytes)